Disallow RegExp "y" flag in ES5 mode
Previously, a regular expression with the "y" flag set would correctly
parse in ES5 mode, though the "y" flag was first introduced in ES6. In
addition, an "s" flag was allowed in both modes, though neither spec
defines it. This change:

- Only permits the "y" flag when `ecmaVersion` is 6.
- Adds tests to verify that "u" and "y" (both new in ES6) fail to parse
  when `ecmaVersion` is less than 6.
- Removes the "s" flag.

Currently the warning message for invalid regular expression flags gives
the location as the start of the regular expression. It may be more
helpful to give the location of the invalid flag, but I left this as-is
for now.

ES5 spec text:
http://www.ecma-international.org/ecma-262/5.1/#sec-15.10.4.1
ES6 spec text:
http://www.ecma-international.org/ecma-262/6.0/#sec-regexpinitialize

Fixes #363