Update to fix optimization pass under Closure Compiler.
Closure Compiler's analyzer (and presumably other analyzers as well) is broken by the mixed struct-like and dict-like usage of the tokentype.types object. The constructed key names of the keyword tokens get mapped to incorrect strings in the rename pass, resulting in a nonfunctional output. Manually specifying the tokentype.types keys fixes the issue.